@@ -42,17 +42,17 @@
     "autobahn": "0.9.9",
     "classnames": "2.2.3",
     "d3": "3.5.12",
-    "esp-js": "1.0.1",
-    "fixed-data-table": "0.6.0",
+    "esp-js": "1.0.3",
+    "fixed-data-table": "0.6.3",
     "font-awesome": "4.5.0",
     "lodash": "3.10.1",
     "moment": "2.13.0",
     "numeral": "1.5.3",
     "nvd3": "1.8.1",
-    "react": "0.14.6",
+    "react": "15.3.0",
     "react-dimensions": "1.1.1",
-    "react-dom": "0.14.6",
-    "react-nvd3": "0.4.1",
+    "react-dom": "15.3.0",
+    "react-nvd3": "0.5.7",
     "react-sparklines": "1.4.2",
     "rx": "4.0.7",
     "ua-parser-js": "0.7.10",
@@ -83,7 +83,6 @@
     "eslint-config-airbnb": "3.1.0",
     "eslint-loader": "1.2.0",
     "eslint-plugin-react": "3.15.0",
-    "esp-js-devtools": "0.0.15",
     "exports-loader": "0.6.3",
     "extract-text-webpack-plugin": "1.0.1",
     "file-loader": "0.8.5",
@@ -102,7 +101,7 @@
     "node-sass": "3.4.2",
     "npm-run-all": "2.1.0",
     "openfin-cli": "1.0.2",
-    "react-addons-test-utils": "0.14.6",
+    "react-addons-test-utils": "15.3.0",
     "react-transform-catch-errors": "1.0.1",
     "react-transform-hmr": "1.0.1",
     "redbox-react": "1.2.0",
